DBDT31647 【临床数据中心3期】ETL管理-作业管理调用ETLhttp执行、生命周期年龄字段类型调整

+++ mediinfo-grus-shujuzx/src/main/java/cn/mediinfo/grus/shujuzx/dto/shengmingzqpzs/SC_ZD_ShengMingZQCreateDto.java
@@ -2,6 +2,9 @@ package cn.mediinfo.grus.shujuzx.dto.shengmingzqpzs;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+import org.apache.poi.hpsf.Decimal;
+
+import java.math.BigDecimal;
 
 @Data
 public class SC_ZD_ShengMingZQCreateDto {
@@ -9,9 +12,9 @@ public class SC_ZD_ShengMingZQCreateDto {
     @Schema(description = "生命周期名称")
     private String shengMingZQMC;
     @Schema(description = "起始年龄")
-    private Integer qiShiNL;
+    private BigDecimal qiShiNL;
     @Schema(description = "结束年龄")
-    private Integer jieShuNL;
+    private BigDecimal jieShuNL;
     @Schema(description = "男性图片")
     private String nanXingTP;
     @Schema(description = "女性图片")
