+++ mediinfo-grus-shujuzx/src/main/java/cn/mediinfo/grus/shujuzx/dto/shengmingzqpzs/SC_ZD_ShengMingZQListDto.java
@@ -3,6 +3,7 @@ package cn.mediinfo.grus.shujuzx.dto.shengmingzqpzs;
 import cn.mediinfo.grus.shujuzx.events.BaseEventDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+import org.apache.poi.hpsf.Decimal;
 
 @Data
 public class SC_ZD_ShengMingZQListDto {
@@ -11,9 +12,9 @@ public class SC_ZD_ShengMingZQListDto {
     @Schema(description = "生命周期名称")
     private String shengMingZQMC;
     @Schema(description = "起始年龄")
-    private Integer qiShiNL;
+    private Decimal qiShiNL;
     @Schema(description = "结束年龄")
-    private Integer jieShuNL;
+    private Decimal jieShuNL;
     @Schema(description = "顺序号")
     private Integer shunXuHao;
 }
